Restructure transport and MQTT
