+++ source/main.c
@@ -6,11 +6,10 @@
 #include <time.h>
 
 #include "core_mqtt.h"
-#include "transport/transport.h"
+#include "transport_wrapper.h"
 #include "utils/clock.h"
 
 static TransportInterface_t transport = { 0 };
-static bool sessionPresent = false;
 static MQTTContext_t mqttContext = { 0 };
 static uint8_t networkBuffer[ 5000U ];
 
@@ -22,6 +21,25 @@ static void mqttEventCallback( MQTTContext_t * mqttContext,
     ( void ) packetInfo;
     ( void ) deserializedInfo;
     printf( "MQTT event callback triggered\n" );
+}
+
+static MQTTStatus_t mqttConnect( char * thingName )
+{
+    MQTTConnectInfo_t connectInfo = { 0 };
+    bool sessionPresent = false;
+    connectInfo.pClientIdentifier = thingName;
+    connectInfo.clientIdentifierLength = ( uint16_t ) sizeof( thingName );
+    connectInfo.pUserName = NULL;
+    connectInfo.userNameLength = 0U;
+    connectInfo.pPassword = NULL;
+    connectInfo.passwordLength = 0U;
+    connectInfo.keepAliveSeconds = 60U;
+    connectInfo.cleanSession = true;
+    return MQTT_Connect( &mqttContext,
+                         &connectInfo,
+                         NULL,
+                         5000U,
+                         &sessionPresent );
 }
 
 int main( int argc, char * argv[] )
@@ -47,6 +65,7 @@ int main( int argc, char * argv[] )
                                          Clock_GetTimeMs,
                                          mqttEventCallback,
                                          &fixedBuffer );
+    assert( mqttResult == MQTTSuccess );
 
     bool result = transport_tlsConnect( certificateFilePath,
                                         privateKeyFilePath,
@@ -54,20 +73,7 @@ int main( int argc, char * argv[] )
                                         endpoint );
     assert( result );
 
-    MQTTConnectInfo_t connectInfo = { 0 };
-    connectInfo.pClientIdentifier = thingName;
-    connectInfo.clientIdentifierLength = ( uint16_t ) sizeof( thingName );
-    connectInfo.pUserName = NULL;
-    connectInfo.userNameLength = 0U;
-    connectInfo.pPassword = NULL;
-    connectInfo.passwordLength = 0U;
-    connectInfo.keepAliveSeconds = 60U;
-    connectInfo.cleanSession = true;
-    mqttResult = MQTT_Connect( &mqttContext,
-                               &connectInfo,
-                               NULL,
-                               5000U,
-                               &sessionPresent );
+    mqttResult = mqttConnect( thingName );
     assert( mqttResult == MQTTSuccess );
     printf( "Successfully connected to IoT Core\n" );
